disable gamepad nav on launcher
Gamepad events are processed on all windows, not just the active one.
This breaks the setup/help modals as the main window can steal focus
or launch an action.

Adding a toggle to enable/disable gamepad nav depending on whether
a modal is open does work for general navigation. Unfortunately the
button down state used to close the modal will still be active on return
to the main window, which immediately re-opens the modal, forcing
the use of a mouse/keyboard to get out of the loop.