Revert "copy FS1 missile behavior change from FSO"
This reverts commit aea7757887615fb5ce3c4aff2f3215a1835b5a39.

After a lot of testing this appears to make a lot of missile launches useless.
In controlled situations it looks good, but in the heat of battle it's really
poor behavior. Regardless of whether or not it's actually more retail-like,
the balance appears shifted to quite heavily favor the AI defenders at the
expense of the player. Additional changes to AI behavior or countermeasure
effectiveness may be required to compensate for the balance shift.

Until such changes are made it's better to just rely on FS2 missile behavior.