fix gamepad nav bug in launcher
Imgui processes gamepad events in all windows all of the time, which doesn't
work out very well when using multiple windows/contexts. So make sure that
we're only processing gamepad events for the active context.

Additionally, a gamepad event as the final action of a closing context
would pass down to the main context. This could cause it to automatically
re-open the setup window (or help)  when closed with the gamepad. So the
main window needs an additional hack to try and avoid those passing events.